+++ src/fullname.ts
@@ -1,21 +1,21 @@
 import { Config } from './config.js';
+import { Validators } from './validator.js';
+import { Nullable, Namon, Title } from './types.js';
 import { NameError, UnknownError } from './error.js';
 import { FirstName, LastName, Name, JsonName } from './name.js';
-import { Nullable, Namon, Title } from './types.js';
-import { Validators } from './validator.js';
 
 /**
  * The core component of this utility.
  *
- * This component is comprised of five entities that make it easy to handle a
+ * This component is composed of five entities that make it easy to handle a
  * full name set: prefix, first name, middle name, last name, and suffix.
- * This class is intended for internal processes. However, it is understandable
+ * It is indeed intended for internal processes. However, it is understandable
  * that it might be needed at some point for additional purposes. For this reason,
  * it's made available.
  *
  * It is recommended to avoid using this class unless it is highly necessary or
- * a custom parser is used for uncommon use cases. This utility tries to cover
- * as many use cases as possible.
+ * a custom parser is used for uncommon use cases although this utility tries to
+ * cover as many use cases as possible.
  *
  * Additionally, an optional configuration can be used to indicate some specific
  * behaviors related to that name handling.
@@ -30,7 +30,7 @@ export class FullName {
 
   /**
    * Creates a full name as it goes
-   * @param options optional configuration for additional features.
+   * @param options settings for additional features.
    */
   constructor(options?: Partial<Config>) {
     this.#config = Config.merge(options);
@@ -67,26 +67,25 @@ export class FullName {
   }
 
   /**
-   * Parses a json name into a full name.
-   * @param json parsable name element
-   * @param config optional configuration for additional features.
+   * Parses a JSON name into a full name.
+   * @param {JsonName} json parsable name element
+   * @param {Config} config for additional features.
    */
   static parse(json: JsonName, config?: Config): FullName {
     try {
-      const fullName = new FullName(config);
-      fullName.setPrefix(json.prefix);
-      fullName.setFirstName(json.firstName);
-      fullName.setMiddleName(json.middleName ?? []);
-      fullName.setLastName(json.lastName);
-      fullName.setSuffix(json.suffix);
-      return fullName;
+      return new FullName(config)
+        .setPrefix(json.prefix)
+        .setFirstName(json.firstName)
+        .setMiddleName(json.middleName ?? [])
+        .setLastName(json.lastName)
+        .setSuffix(json.suffix);
     } catch (error) {
       if (error instanceof NameError) throw error;
 
       throw new UnknownError({
         source: Object.values(json).join(' '),
         message: 'could not parse JSON content',
-        error: error instanceof Error ? error : new Error(String(error)),
+        origin: error instanceof Error ? error : new Error(String(error)),
       });
     }
   }
@@ -114,7 +113,7 @@ export class FullName {
   setMiddleName(names: string[] | Name[]): FullName {
     if (!Array.isArray(names)) return this;
     if (!this.#config.bypass) Validators.middleName.validate(names);
-    this.#middleName = (names as Array<string | Name>).map((name) => (name instanceof Name ? name : Name.middle(name)));
+    this.#middleName = (names as Array<string | Name>).map((n) => (n instanceof Name ? n : Name.middle(n)));
     return this;
   }
 
@@ -125,9 +124,7 @@ export class FullName {
     return this;
   }
 
-  /**
-   * Returns true if a namon has been set.
-   */
+  /** Returns true if a namon has been set. */
   has(namon: Namon): boolean {
     if (namon.equal(Namon.PREFIX)) return !!this.#prefix;
     if (namon.equal(Namon.SUFFIX)) return !!this.#suffix;